Fixed design-time region display
I accidentally jacked up design time region displays. The designer was
dutifully calling the Map constructor, which obediently gets the map's
adjacencies. The design-time data was set up to throw an exception,
which I guess the designer swallows?

At any rate, it seems this manifested as the map xml not loading
properly, which was also done in the constructor.